pyproject.toml
[tool.poetry]
name = "devgrid-pychallenge"
version = "0.1.0"
description = ""
authors = ["Tales Araujo <talesaraujodossantos@gmail.com>"]
readme = "README.md"
packages = [{include = "devgrid_pychallenge"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
fastapi = "^0.103.2"
httpx = "^0.25.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.16.1"
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
